MD5 conversion class
This is the class I worked on that will us to create an object that has 2 fields: plaintext and md5hash. When we create an object of this type, we will input just the plaintext, and it will determine the hash for us. (as i am writing this, i realize there is some editing to do for this class that will change the fields)